requirements.txt
selenium==4.3.0
python-dotenv